requirements.txt
python-dotenv>=1.1.1
fastapi>=0.117.1
sqlalchemy>=2.0.43
uvicorn[standard]>=0.37.0
email-validator>=2.3.0
bcrypt>=4.3.0
python-jose[cryptography]>=3.5.0
python-multipart
asyncpg>=0.30.0
psycopg2-binary>=2.9.10

